Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/close/close-basic-expected.txt


FAIL WebSockets: close() assert_equals: expected 2 but got 3
